LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-fallback-invalidation-expected.txt

X

FAIL Correct invalidation when fallbacks match the old style assert_equals: expected 40 but got 0
